Reimplement case handling in LookupString to match XCB
The primary change here is that we do not use the unicode package to
operate on strings. Instead, we implement XCB's xcb_convert_case
function, which is a truer implementation of the specification, handling
only those keysyms it should.

In itself, this isn't a very useful change. It does, however, allow us
to return keysyms, instead of strings, from interpretSymList. This is
necessary to add support for handling numlock, which has to check
whether the keysym is in specific ranges.